Moved errors_logged() to new utils file
